pyproject.toml
[tool.poetry]
name = "api_pytest"
version = "1.0.0"
description = ""
authors = ["BenjaminYing <fzuyingming@163.com>"]

[tool.poetry.dependencies]
python = "3.6.9"
pytest = "^5.4.2"
requests = "^2.23.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
